Week6/requirements.txt
uvicorn[standard]
fastapi
jinja2
python-multipart
starlette
itsdangerous
mysql-connector-python 
python-dotenv